# fix fastpaths for materializing transposes / adjoints of sparse matrices