--- a/stdlib/SparseArrays/test/sparse.jl
+++ b/stdlib/SparseArrays/test/sparse.jl
@@ -2242,4 +2242,10 @@ _length_or_count_or_five(x) = length(x)
     end
 end
 
+@testset "sparse transpose adjoint" begin
+    A = sprand(10, 10, 0.75)
+    @test A' == SparseMatrixCSC(A')
+    @test transpose(A) == SparseMatrixCSC(transpose(A))
+end
+
 end # module