Setup:

```jl
using LinearAlgebra
using SparseArrays
using BenchmarkTools
n = 1000
A = sparse(1.0I,n,n);
B = sparse(1.0I,n,n);

```

Before PR:

```jl
julia> @btime $A + $B';
  5.307 ms (28 allocations: 112.11 KiB)

julia> @btime $A + $(B');
  5.305 ms (27 allocations: 112.09 KiB)
```


After PR:

```jl
julia> @btime $A + $B';
  13.897 μs (10 allocations: 63.52 KiB)

julia> @btime $A + $(B');
  13.852 μs (9 allocations: 63.50 KiB)
```




Fixes https://github.com/JuliaLang/julia/issues/28012